src/pyproject.toml
[tool.poetry]
name = "delivery"
version = "0.1.0"
description = ""
authors = ["undergroundenemy616 <undergroundenemy616@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
ruff = "^0.5.0"
isort = "^5.13.2"
black = "^24.4.2"
autoflake = "^2.3.1"
mypy = "^1.10.1"
pre-commit = "^3.7.1"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
greenlet = "^3.0.3"
pytest-asyncio = "^0.23.7"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
redis = "^5.0.7"
celery = "^5.4.0"
punq = "^0.7.0"
grpcio = "^1.64.1"
grpcio-tools = "^1.64.1"
faststream = "^0.5.14"
aiokafka = "0.10.0"
pydeps = "^1.12.20"
networkx = "^3.3"
matplotlib = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.autoflake]
remove-all-unused-imports = true
ignore-init-module-imports = true
expand-star-imports = true
remove-duplicate-keys = true

[tool.black]
line-length = 120
target-version = ['py311']
exclude = '''
(
  \.eggs
  |\.git
  |\.hg
  |\.mypy_cache
  |\.nox
  |\.tox
  |\.venv
  |\.pyc
  |\.pyi
  |_build
  |buck-out
  |build
  |dist
  |venv
)
'''

[tool.isort]
line_length = 120
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_output = 3
profile = "black"
src_paths = "."
lines_after_imports = 2


[tool.ruff]
line-length = 120
target-version = "py311"
fix = true
lint.select = [
    "F",   # pyflakes
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "C90", # mccabe
    "N",   # pep8-naming
    "C4",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "BLE", # flake8-blind-except
    "T20", # flake8-print
    "Q",   # flake8-quotes
]
exclude = [
    "eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".tox",
    ".venv",
    ".pyc",
    ".pyi",
    "_build",
    "buck-out",
    "build",
    "dist",
    "migrations",
    "proto",
]

[tool.ruff.lint.mccabe]
# Настройка максимальной сложности
max-complexity = 5


[tool.mypy]
explicit_package_bases = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = '''
(
  \.eggs
  |\.git
  |\.hg
  |\.mypy_cache
  |\.nox
  |\.tox
  |\.venv
  |\.pyc
  |\.pyi
  |_build
  |buck-out
  |build
  |dist
  |venv
)
'''